painting a wall
Input: Two arguments.

The required length of the wall that should be painted, as integer.
A list of the operations that contains the range (inclusive) as a list of two integers.
Output: The minimum number of operations. If you cannot paint enough of the length with the given operations, return -1.